NC_VIT/requirements.txt
torch==2.0.1 
torchvision==0.15.2 
numpy==1.25.0 
matplotlib==3.7.1 
scipy==1.11.1   
scikit-learn==1.3.0  
Pillow==9.5.0 
tqdm==4.65.0
